@@ -5,7 +5,6 @@
 # SPDX-License-Identifier: LGPL-3.0-or-later
 """Restriction of a DofMap to a list of active degrees of freedom."""
 
-import typing
 
 import dolfinx.cpp as dcpp
 import dolfinx.fem
@@ -20,7 +19,7 @@ class DofMapRestriction(mcpp.fem.DofMapRestriction):  # type: ignore[misc, no-an
 
     def __init__(  # type: ignore[no-any-unimported]
         self,
-        dofmap: typing.Union[dcpp.fem.DofMap, dolfinx.fem.DofMap],
+        dofmap: dcpp.fem.DofMap | dolfinx.fem.DofMap,
         restriction: npt.NDArray[np.int32]
     ) -> None:
         # Extract cpp dofmap
